+++ src/predicates/mod.rs
@@ -1,6 +1,8 @@
 mod expansion;
 mod orient2d;
+mod orient3d;
 mod sign;
 
 pub use orient2d::orient2d;
+pub use orient3d::orient3d;
 pub use sign::{Orientation, Sign};